refactor: extract generatorDispatch + factBreakout to forge/ (tier-2, #30)
Move the byte-identical generatorDispatch, writeUnder, and factBreakout
functions from ai-forge/ and saas-forge/ into forge/generators.mjs and
forge/breakouts.mjs. Both products re-export from the shared module so
downstream import paths are unchanged.

Product-specific code stays product-side:
- ai-forge: makePatternGenerators, runPatternBreakouts
- saas-forge: makeDemoGenerators, runTeamBreakouts (incl. fight logs)

Adds forge/scripts/test-boundaries.mjs — a CI grep guard that asserts
neither product imports from the other.
