+++ forge/package.json
@@ -5,8 +5,8 @@
   "type": "module",
   "description": "The adversarial-convergence doctrine as an engine: ratchet runs (Styx rule, blocker respec, contract closure, verify-failure banking, evidence-digest approvals) and fixed-point drivers, extracted from the gate-PASSED live runs.",
   "scripts": {
-    "check": "node --check ratchet.mjs && node --check driver.mjs && node --check manifest.mjs && node --check claims.mjs && node --check operator.mjs && node --check scripts/test-ratchet.mjs && node --check scripts/test-driver.mjs && node --check scripts/test-manifest.mjs && node --check scripts/test-claims.mjs && node --check scripts/test-operator.mjs",
-    "test": "npm run check && node scripts/test-ratchet.mjs && node scripts/test-driver.mjs && node scripts/test-manifest.mjs && node scripts/test-claims.mjs && node scripts/test-operator.mjs"
+    "check": "node --check ratchet.mjs && node --check driver.mjs && node --check manifest.mjs && node --check claims.mjs && node --check operator.mjs && node --check generators.mjs && node --check breakouts.mjs && node --check scripts/test-ratchet.mjs && node --check scripts/test-driver.mjs && node --check scripts/test-manifest.mjs && node --check scripts/test-claims.mjs && node --check scripts/test-operator.mjs && node --check scripts/test-boundaries.mjs",
+    "test": "npm run check && node scripts/test-ratchet.mjs && node scripts/test-driver.mjs && node scripts/test-manifest.mjs && node scripts/test-claims.mjs && node scripts/test-operator.mjs && node scripts/test-boundaries.mjs"
   },
   "engines": {
     "node": ">=18"
